+++ lib/src/matrix/matrix_cache_keys.dart
@@ -1,11 +1,19 @@
-const meansByRowsKey = 'means_by_rows';
-const meansByColumnsKey = 'means_by_columns';
-const deviationByRowsKey = 'deviation_by_rows';
-const deviationByColumnsKey = 'deviation_by_columns';
+const matrixMeansByRowsKey = 'means_by_rows';
+const matrixMeansByColumnsKey = 'means_by_columns';
+const matrixDeviationByRowsKey = 'deviation_by_rows';
+const matrixDeviationByColumnsKey = 'deviation_by_columns';
+const matrixPowKey = 'pow';
+const matrixExpKey = 'exp';
+const matrixSumKey = 'sum';
+const matrixProdKey = 'prod';
 
 final matrixCacheKeys = Set<String>.from(<String>[
-  meansByRowsKey,
-  meansByColumnsKey,
-  deviationByRowsKey,
-  deviationByColumnsKey,
+  matrixMeansByRowsKey,
+  matrixMeansByColumnsKey,
+  matrixDeviationByRowsKey,
+  matrixDeviationByColumnsKey,
+  matrixPowKey,
+  matrixExpKey,
+  matrixSumKey,
+  matrixProdKey,
 ]);