+++ test/integration_test/vector/methods/prod/prod_test.dart
@@ -1,0 +1,8 @@
+import 'package:ml_linalg/dtype.dart';
+
+import 'prod_test_group_factory.dart';
+
+void main() {
+  vectorProdTestGroupFactory(DType.float32);
+  vectorProdTestGroupFactory(DType.float64);
+}